run-3dgs/requirements.txt
fastapi
uvicorn
google-cloud-logging